+++ src/collective/polls/Extensions/Install.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 
 from collective.polls.config import PROJECTNAME
-from Products.CMFCore.utils import getToolByName
+from plone import api
 
 
 def uninstall(portal):
     profile = 'profile-%s:uninstall' % PROJECTNAME
-    setup_tool = getToolByName(portal, 'portal_setup')
+    setup_tool = api.portal.get_tool(name='portal_setup')
     setup_tool.runAllImportStepsFromProfile(profile)
     return "Ran all uninstall steps."